Image gallery is working.
I created the image gallery implementation.

All the system frags have to do is provide a list
of image resources they want to display in their gallery.

I also added a third pary library for image zooming.
When the user clicks on an image they're given a full screen
view of the image with the ability to zoom in on it.

I also found some image resources for the different systems.
I'm currently using them as a placeholder for images in the sys frags.